Make pybdsf/pybdsm optional
The interactive BDSF shell(s) `pybdsf` and `pybdsm` depend on IPython,
which we don't want to install by default. We therefore created an
extras called "ishell", which will install IPython when requested.

NOTE: Due to limitations of the current setuptools back-end, the scripts
`pybdsf` and `pybdsm` are _always_ installed, also if the user did not
request the extra "ishell". I don't know how to avoid this.