Fix py38 build issue (#174)
* Fix broken PyBDSF shell

As a result of the recent change where the `pybdsf` script was converted into a module with a `main` function that was defined as `entry_point` in `setup.py`, the interactive shell no longer worked properly. You could start it, but you could not execute any of the commands. This commit fixes this. It has been tested manually on FITS-file available in the test directory, and it seems to work fine.