Switch to `manylinux2014` for building binary wheels (#191)
* Use manylinux2014 for building wheels

Package `manylinux2010` is end-of-life, so we needed to switch to `manylinux2014`.

* Refactor script and docker files for building wheels

The different `wheel*.docker` files used to build the wheels for the different python version contained a lot a code duplication.
There's now a single `py_wheel.docker` that can build wheels for multiple python version. The build of these wheels is still controlled by `build.sh`, which was also rewritten.