Fix build issue with Python 3.11
The Python 3.11 build failed, due to some obscure error generated by `f2py`.
The reason this problem only arises with Python 3.11 is that it requires NumPy 1.24, which introduced this bug(?).
I'm not sure why it works, but if I reorder the two source code file names in `setup.py`, the build succeeds.

Fixes #189