Use scikit-build (#204)
Use `scikit-build` to build the software.

When using `scikit-build`, the build is driven by CMake and the contents of the declarative `CMakeLists.txt` file. As a result, we can get rid of all the extra code in `setup.py`, which is rather complicated, hard to maintain, and soon deprecated (because it depends on `distutils`). The only thing remaining in `setup.py` is a single function call to `skbuild.setup`, which mimics the behavior of `setuptools.setup`.