Check exit code of each joined process (#255)
We need to check the exit code of each joined process in the `run_tasks()` function in `multi_proc.py`. A non-zero exit status (especially a negative exit status) indicates that an error occurred, and it is possible that the output queue (`out_q`) will be empty. Not handling this situation properly, will cause `out_q.get()` (a few lines further) to hang.